pyproject.toml
[project]
name = "photos-wrapped"
version = "0.1.0"
description = "Add your description here"
authors = [
    { name = "Dustin Knopoff", email = "dustin@knopoff.dev" }
]
dependencies = [
    "osxphotos>=0.68.6",
    "fastapi[standard]>=0.115.5",
    "black>=24.4.2",
    "litequeue>=0.8",
    "sqlalchemy>=2.0.31",
    "polling>=0.3.2",
    "jinja2>=3.1.4",
    "pillow>=10.4.0",
    "pillow-heif>=0.17.0",
    "pillow-avif-plugin>=1.4.3",
    "sentence-transformers>=3.0.1",
    "pydantic>=2.8.2",
    "pandas>=2.2.3",
    "scikit-learn>=1.5.2",
    "numpy>=1.99.0",
    "sqlmodel>=0.0.22",
]
readme = "README.md"
requires-python = ">= 3.10"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.rye]
managed = true
dev-dependencies = [
    "ipdb>=0.13.13",
    "alembic>=1.14.0",
]

[tool.hatch.metadata]
allow-direct-references = true

[tool.hatch.build.targets.wheel]
packages = ["src/photos_wrapped"]
